refactor: rename and simplify user validation for group members
Remove references to 'verified' concept from group member validation.
The validation now simply checks that a user exists before being
assigned to any role, not just elevated roles.

- Rename VerifiedForElevatedRoles to UserExistsForRole
- Simplify validation to check all roles, not just owner/organizer
- Update module documentation to reflect actual behavior
